+++ src/API/PostAPI.js
@@ -1,4 +1,12 @@
-export default function PostsAPI()
-{
-    return (fetch("https://jsonplaceholder.typicode.com/posts"));
-}
+export const PostsAPI = async () => {
+
+    try {
+      const response = await fetch("https://jsonplaceholder.typicode.com/posts"); 
+      const dataFetched = await response.json();
+      return dataFetched;
+    } catch (error) {
+      console.error("Error fetching data:", error);
+      return null;
+    }
+  };
+
