+++ src/context/Posts/PostStateContext.js
@@ -1,27 +1,18 @@
-//Builtin Libraries Imports
 import React, { useEffect, useState } from "react";
 
-//Custom File Imports
 import PostContext from "./PostContext.js";
-import PostsAPI from "../../api/PostAPI.js";
-import {
-  initializePostDataOnLocalStorage,
-  setPostDataOnLocalStorage,
-} from "../../utils/Posts/postsLocalStorage.js";
+import {PostsAPI} from "../../api/PostAPI.js";
+import {initializePostDataOnLocalStorage,setPostDataOnLocalStorage,} from "../../utils/Posts/postsLocalStorage.js";
 
 export default function PostStateContext(props) {
-  // Check if data exists in localStorage,
-  //if not, initialize with an empty array
   const initialData = initializePostDataOnLocalStorage();
-  // Creating State Variable to Fetch Post Data
   const [postData, setPostData] = useState(initialData);
-  // Using UseEffect Hook To Fetch Data and
-  // Update the React Component Only Once
+
   useEffect(() => {
     async function fetchAndSetData() {
       // Fetch data only if postData is empty
       if (postData.length === 0) {
-        const dataFetched = await fetchData();
+        const dataFetched = await PostsAPI();
         if (dataFetched) {
           setPostData(dataFetched);
 
@@ -30,14 +21,11 @@ export default function PostStateContext(props) {
         }
       }
     }
-
     fetchAndSetData();
-  }, [postData]); // Add postData as a dependency to prevent unnecessary re-fetching
+  }, [postData]);
 
-  // updates the PostData and postData inlocalStorage
   const updatePost = (data) => {
     setPostData(data);
-    // Save updated data to localStorage
     setPostDataOnLocalStorage(data);
   };
 
@@ -48,13 +36,4 @@ export default function PostStateContext(props) {
   );
 }
 
-async function fetchData() {
-  try {
-    const response = await PostsAPI();
-    const dataFetched = await response.json();
-    return dataFetched;
-  } catch (error) {
-    console.error("Error fetching data:", error);
-    return null;
-  }
-}
+
