+++ src/pages/Auth/login.js
@@ -1,0 +1,8 @@
+import Auth from './auth';
+
+import 'bootstrap/dist/css/bootstrap.min.css';
+
+export default function Login(props)
+{
+    return (<Auth flag={false}/>)
+}
