+++ src/pages/Auth/register.js
@@ -1,0 +1,8 @@
+import Auth from './auth';
+
+import 'bootstrap/dist/css/bootstrap.min.css';
+
+export default function Register(props)
+{
+    return (<Auth flag={true}/>);
+}
