+++ src/pages/Posts/editBody.js
@@ -1,4 +1,4 @@
-import ComposePost from "../components/posts/composePost";
+import ComposePost from "../../components/posts/composePost";
 export default function EditBody()
 {
     return (<ComposePost flag={false}/>);
