+++ src/pages/Posts/writeBody.js
@@ -1,4 +1,4 @@
-import ComposePost from "../components/posts/composePost";
+import ComposePost from "../../components/posts/composePost";
 export default function WriteBody()
 {
     return (<ComposePost flag={true}/>);
