+++ src/styles/login.css
@@ -36,6 +36,7 @@
     color:grey;
     margin-left:2.5vh;
     margin-bottom:3vh;  
+    text-decoration: underline;
 }
 
 .div-wrapper
