+++ src/utils/Posts/postsLocalStorage.js
@@ -1,12 +1,12 @@
-export function initializePostDataOnLocalStorage()
+export const initializePostDataOnLocalStorage=()=>
 {
     return (localStorage.getItem("postData")
     ? JSON.parse(localStorage.getItem("postData"))
     : []);
 }
 
-export function setPostDataOnLocalStorage(data)
+export const setPostDataOnLocalStorage=(data)=>
 {
-   return  (localStorage.setItem("postData", JSON.stringify(data)));
+  localStorage.setItem("postData", JSON.stringify(data));
 }
